+++ CHANGES.rst
@@ -5,7 +5,7 @@ Changelog of lizard-graph
 0.16.2 (unreleased)
 -------------------
 
-- Nothing changed yet.
+- Added option unit-as-y-label, which can be set to True.
 
 
 0.16.1 (2012-04-19)